+++ locals-tags.tf
@@ -1,0 +1,7 @@
+locals {
+  bastion_tags = var.default_tags_enabled ? {
+    env    = var.environment
+    stack  = var.stack
+    module = "bastion"
+  } : {}
+}