+++ variables.tf
@@ -107,21 +107,3 @@ variable "storage_os_disk_size_gb" {
   description = "Specifies the size of the OS Disk in gigabytes"
   type        = string
 }
-
-variable "bastion_extra_tags" {
-  description = "Additional tags to associate with your bastion instance."
-  type        = map(string)
-  default     = {}
-}
-
-variable "ani_extra_tags" {
-  description = "Additional tags to associate with your network interface."
-  type        = map(string)
-  default     = {}
-}
-
-variable "pubip_extra_tags" {
-  description = "Additional tags to associate with your public ip."
-  type        = map(string)
-  default     = {}
-}